updating protein tests
